requirements.txt
face_recognition==1.3.0
Pillow==11.0.0
tqdm==4.67.0
